+++ src/main/org/codehaus/groovy/runtime/dgmimpl/arrays/ObjectArrayPutAtMetaMethod.java
@@ -27,7 +27,7 @@ import org.codehaus.groovy.runtime.typehandling.DefaultTypeTransformation;
 public class ObjectArrayPutAtMetaMethod extends ArrayPutAtMetaMethod {
     private static final CachedClass OBJECT_CLASS = ReflectionCache.getCachedClass(Object.class);
     private static final CachedClass OBJECT_ARR_CLASS = ReflectionCache.OBJECT_ARRAY_CLASS;
-    private static final CachedClass [] PARAM_CLASS_ARR = new CachedClass[] {INTEGER_CLASS, OBJECT_CLASS};
+    private static final CachedClass[] PARAM_CLASS_ARR = new CachedClass[]{INTEGER_CLASS, OBJECT_CLASS};
 
     public ObjectArrayPutAtMetaMethod() {
         parameterTypes = PARAM_CLASS_ARR;
@@ -39,32 +39,32 @@ public class ObjectArrayPutAtMetaMethod extends ArrayPutAtMetaMethod {
 
     public Object invoke(Object object, Object[] arguments) {
         final Object[] objects = (Object[]) object;
-    	final int index = normaliseIndex(((Integer) arguments[0]).intValue(), objects.length);
+        final int index = normaliseIndex((Integer) arguments[0], objects.length);
         objects[index] = adjustNewValue(objects, arguments[1]);
         return null;
     }
-    
+
     private static Object adjustNewValue(Object[] objects, Object newValue) {
-    	Class arrayComponentClass = objects.getClass().getComponentType();
-    	Object adjustedNewVal = newValue;
+        Class arrayComponentClass = objects.getClass().getComponentType();
+        Object adjustedNewVal = newValue;
         if (newValue instanceof Number) {
-        	if (!arrayComponentClass.equals(newValue.getClass())) {
-        		adjustedNewVal = DefaultTypeTransformation.castToType(newValue, arrayComponentClass);
-        	}
+            if (!arrayComponentClass.equals(newValue.getClass())) {
+                adjustedNewVal = DefaultTypeTransformation.castToType(newValue, arrayComponentClass);
+            }
         } else if (Character.class.isAssignableFrom(arrayComponentClass)) {
-        	adjustedNewVal = DefaultTypeTransformation.getCharFromSizeOneString(newValue);
+            adjustedNewVal = DefaultTypeTransformation.getCharFromSizeOneString(newValue);
         } else if (Number.class.isAssignableFrom(arrayComponentClass)) {
-        	if(newValue instanceof Character || newValue instanceof String || newValue instanceof GString) {
-        		Character ch = DefaultTypeTransformation.getCharFromSizeOneString(newValue);
-        		adjustedNewVal = DefaultTypeTransformation.castToType(ch, arrayComponentClass);
-        	}
+            if (newValue instanceof Character || newValue instanceof String || newValue instanceof GString) {
+                Character ch = DefaultTypeTransformation.getCharFromSizeOneString(newValue);
+                adjustedNewVal = DefaultTypeTransformation.castToType(ch, arrayComponentClass);
+            }
         }
         return adjustedNewVal;
     }
 
     public CallSite createPojoCallSite(CallSite site, MetaClassImpl metaClass, MetaMethod metaMethod, Class[] params, Object receiver, Object[] args) {
-        if (!(args [0] instanceof Integer))
-          return PojoMetaMethodSite.createNonAwareCallSite(site, metaClass, metaMethod, params, args);
+        if (!(args[0] instanceof Integer))
+            return PojoMetaMethodSite.createNonAwareCallSite(site, metaClass, metaMethod, params, args);
         else
             return new MyPojoMetaMethodSite(site, metaClass, metaMethod, params);
     }
@@ -78,15 +78,15 @@ public class ObjectArrayPutAtMetaMethod extends ArrayPutAtMetaMethod {
             if (checkPojoMetaClass()) {
                 try {
                     final Object[] objects = (Object[]) receiver;
-                    objects[normaliseIndex(((Integer) arg1).intValue(), objects.length)] = adjustNewValue(objects, arg2);
+                    objects[normaliseIndex((Integer) arg1, objects.length)] = adjustNewValue(objects, arg2);
                     return null;
                 }
                 catch (ClassCastException e) {
                     if ((receiver instanceof Object[]) && (arg1 instanceof Integer))
-                      throw e;
+                        throw e;
                 }
             }
-            return super.call(receiver,arg1,arg2);
+            return super.call(receiver, arg1, arg2);
         }
     }
 }
